+++ CHANGES.md
@@ -5,3 +5,12 @@ Major Changes by Revision
 -----
 
 None: Initial release
+
+0.9.1
+-----
+
+- Crawler:
+    - Code somewhat tidier
+    - Supports multiple threads
+    - Added an accept option
+    - Removed the depth option